refactor(embed): インラインテストを各 module/tests.rs に分離
